+++ docs/guide/README.md
@@ -4,20 +4,6 @@
 
 vue-chart-3 is rewrite of `vue-chartjs` for [Chart.js 3](https://www.chartjs.org/) for Vue 2 & 3, but written in Typescript and Vue Composition API.
 
-::: warning
-
-# Breaking changes and new versions numbers
-
-In previous versions, `vue-demi` was used. It worked really well, but as the project advanced, they were a lot of problems with vue-related typescript definitions, tests and conflict between dependencies.
-
-This new system will keep `vue-chart-3` working for both Vue 2 & 3 with designated versions (`2.x` and `3.x` respectively), but each one designed for their specific Vue version.
-
-The usage of components of hooks is unchanged for both versions.
-
-Code wise, there is not big breaking changes. But the Vue 3 version will have improved type checking for components template (with Volar extension).
-
-:::
-
 # Installation
 
 ## For Vue 3
