+++ src/core/index.ts
@@ -1,0 +1,2 @@
+export * from './component.builder';
+export * from './component.types';
