+++ src/exports/hooks.exports.ts
@@ -1,0 +1,57 @@
+import { defineChartHook } from '../hooks';
+
+/**
+ * {@link - https://vue-chart-3.netlify.app/components/hooks/#guide}
+ *
+ * ❗️ Still in beta, API can change in future updates
+ */
+export const useDoughnutChart = defineChartHook('doughnut');
+
+/**
+ * {@link - https://vue-chart-3.netlify.app/components/hooks/#guide}
+ *
+ * ❗️ Still in beta, API can change in future updates
+ */
+export const useBarChart = defineChartHook('bar');
+
+/**
+ * {@link - https://vue-chart-3.netlify.app/components/hooks/#guide}
+ *
+ * ❗️ Still in beta, API can change in future updates
+ */
+export const useLineChart = defineChartHook('line');
+
+/**
+ * {@link - https://vue-chart-3.netlify.app/components/hooks/#guide}
+ *
+ * ❗️ Still in beta, API can change in future updates
+ */
+export const usePieChart = defineChartHook('pie');
+
+/**
+ * {@link - https://vue-chart-3.netlify.app/components/hooks/#guide}
+ *
+ * ❗️ Still in beta, API can change in future updates
+ */
+export const usePolarAreaChart = defineChartHook('polarArea');
+
+/**
+ * {@link - https://vue-chart-3.netlify.app/components/hooks/#guide}
+ *
+ * ❗️ Still in beta, API can change in future updates
+ */
+export const useRadarChart = defineChartHook('radar');
+
+/**
+ * {@link - https://vue-chart-3.netlify.app/components/hooks/#guide}
+ *
+ * ❗️ Still in beta, API can change in future updates
+ */
+export const useBubbleChart = defineChartHook('bubble');
+
+/**
+ * {@link - https://vue-chart-3.netlify.app/components/hooks/#guide}
+ *
+ * ❗️ Still in beta, API can change in future updates
+ */
+export const useScatterChart = defineChartHook('scatter');
