+++ src/exports/index.ts
@@ -1,0 +1,2 @@
+export * from './component.exports';
+export * from './hooks.exports';
