+++ src/index.ts
@@ -1,15 +1,5 @@
-import { defineChartComponent } from './components';
+export { defineChartComponent } from './core';
+export { defineChartHook } from './hooks';
+export * from './exports';
 
-export { defineChartComponent } from './components';
 export type { ExtractComponentData, ExtractComponentProps } from './utils';
-
-export * from './hooks';
-
-export const BarChart = defineChartComponent('bar-chart', 'bar');
-export const DoughnutChart = defineChartComponent('doughnut-chart', 'doughnut');
-export const LineChart = defineChartComponent('line-chart', 'line');
-export const PieChart = defineChartComponent('pie-chart', 'pie');
-export const PolarAreaChart = defineChartComponent('polar-chart', 'polarArea');
-export const RadarChart = defineChartComponent('radar-chart', 'radar');
-export const BubbleChart = defineChartComponent('bubble-chart', 'bubble');
-export const ScatterChart = defineChartComponent('scatter-chart', 'scatter');
