+++ src/misc/vue.types.ts
