+++ src/utils/common.utils.ts
@@ -1,0 +1,4 @@
+import camelCase from 'lodash-es/camelCase';
+import startCase from 'lodash-es/startCase';
+
+export const pascalCase = (str: string) => startCase(camelCase(str)).replace(/ /g, '');
